Bug 1211659 - Use mozci 0.15.1 which fixes a backfilling issue with coalesced builds
The issue we are facing is that when a backfill is requested and we have coalesced builds,
instead of mozci thinking that we have no build for a push we think that there is a build,
however, what we use is the build with which we got coalesced.

In other words, we were running tests against the wrong build (the one for the most
recent push).

This release of mozci fixes this issue.