Bug 1214625 - Filling a revision on m-a/m-b fails
When a user requested on Treeherder a revision to fill all missing jobs,
we would fail once we would try to schedule the talos opt builders (bug
1149514).

Mozci's version 0.16.0 fixes this issue by not considering talos opt
builders (when there is an equivalent pgo builder) to not be considered
as part of all available builders for a revision.

This will make us only iterate over valid builders.

We're also bumping the version of the taskcluster client since it does
not have the maximum upper limit restriction for the requests package.