Bug 1253700 - Use new version of mozci which fixes invalid graphs
In the previous version of mozci, we were submitting invalid graphs.
These graphs contained metadata in the individual tasks pointing to
users as ffxbld (which is an automation push), however, TaskCluster
expects an email address and fails.

The new version of mozci, uses the graph's metadata in the tasks which
is valid.