Fix previous release and other changes
It seems that the data from the normalized build exchange is rather
different.
In this change we fix how we determine the repository and the revision
for it.

We also make it easier to determine which environment variables you
still have to set.

We also make Message messages full fize instead of capping at 150
characters to improve debuggability.