Pulse messages that are not processed should not show up on Treeherder
We've some pulse messages that we filter out (ignore) and we don't want
to submit a 'Sch' job to Treeherder in those cases.

This change adds to each of the handlers an ignore() method that allow
us to quickly know if a message will be processed or not.

In those cases we don't submit a Treeherder 'Sch' job.

There are also other significant changes in here:
* all handlers now return an exit code
  * such exit code is used to set the colour of the 'Sch' job
* we log more info for each request
* all handlers now have the same function name to process a pulse message
  * this will be important in the future when wanting to normalize all
  * handlers