Lower 'allowing users' message from error to warning
We were using error() as the alerting level to let me know which users
are being allowed to use the system where their email address does not
end with @mozilla.com or they have not acting on a push they did not
push.

I'm changing the message to warning() since I've set up another
different alert on Papertrail to send me emails only to me.

The error() and Traceback alerting is now going to more users than just
me.