Bug 1355820 - Acknowledge all Pulse messages instead of hoping the retry would work
The original design of pulse_actions was to not acknowledge messages
when something would go wrong in order that it would get retried.

There were too mistakes in that approach:
1) The majority of the failures are permanent
2) The retrying is not until the next time the dyno restarts

From my experience, once something would fail it would get stuck in the
queue until I would empty the queue or find a way to fix the edge case.

I'm going to move all acknowledgements to the top level and not leave
any messages unacknowledged.